finance-dash/requirements.txt
numpy
pandas
matplotlib
seaborn
streamlit
plotly
scipy
python-bcb
xlrd